Clean up AgpHandler access and deposit instance in extras
