+++ slack-plugin/src/main/kotlin/slack/gradle/settings/SgpSettingsPlugin.kt
@@ -17,15 +17,37 @@ package slack.gradle.settings
 
 import org.gradle.api.Plugin
 import org.gradle.api.initialization.Settings
-import slack.gradle.AgpHandlers
+import org.gradle.api.plugins.ExtensionAware
+import slack.gradle.agp.AgpHandler
+import slack.gradle.agp.AgpHandlerFactory
+import slack.gradle.settings.SgpSettingsPlugin.Companion.AGP_HANDLER_EXTENSION_KEY
 
 public class SgpSettingsPlugin : Plugin<Settings> {
   override fun apply(settings: Settings) {
-    AgpHandlers.createSettingsHandler().apply(settings)
-    val handler = AgpHandlers.createHandler()
+    val agpHandlerFactory = AgpHandlerFactory.load()
+    agpHandlerFactory.createSettingsHandler().apply(settings)
+    val handler = agpHandlerFactory.createHandler()
 
     // as Project objects cannot query for the Settings object (and its extensions), we
     // deposit the handler instance into each project using the extra Properties.
-    settings.gradle.beforeProject { extensions.extraProperties["_sgp_agphandler"] = handler }
+    settings.gradle.beforeProject {
+      extensions.extraProperties[AGP_HANDLER_EXTENSION_KEY] = handler
+    }
+  }
+
+  internal companion object {
+    const val AGP_HANDLER_EXTENSION_KEY: String = "_sgp_agphandler"
+  }
+}
+
+/** Retrieves the cached [AgpHandler] from extras. This is deposited here by [SgpSettingsPlugin]. */
+internal fun ExtensionAware.agpHandler(): AgpHandler {
+  val properties = extensions.extraProperties
+  if (properties.has(AGP_HANDLER_EXTENSION_KEY)) {
+    return properties.get(AGP_HANDLER_EXTENSION_KEY) as AgpHandler
+  } else {
+    error(
+      "No AgpHandler found in extraProperties! Did you forget to apply the SGP settings plugin?"
+    )
   }
 }
